Fix `Style/OneClassPerFile` cop - request lib
